requirements.txt
PyMuPDF==1.23.9
faiss-cpu==1.7.4
sentence-transformers==2.2.2
requests==2.31.0
ollama==3.2.0
